+++ db/db.py
@@ -2,7 +2,6 @@ import pathlib
 
 from playhouse.sqlite_ext import SqliteDatabase
 
-from .module import CategoryModule, NovelModule
 
 pathlib.Path("data").mkdir(parents=True, exist_ok=True)
 
@@ -16,21 +15,3 @@ DATABASE = SqliteDatabase(
         "synchronous": 0,
     },
 )
-
-
-class DataBase:
-    NovelModule: NovelModule
-    CategoryModule: CategoryModule
-
-    def __init__(self) -> None:
-        self.database = DATABASE
-
-        self.connect()
-        self.database.create_tables([NovelModule, CategoryModule], safe=True)
-
-    def connect(self):
-        self.database.connect()
-        return self
-
-    def close(self):
-        self.database.close()
